mangaparser/requirements.txt
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
Pillow==10.0.1
pillow-avif-plugin==1.3.2
tqdm==4.66.1